src/test/java/com/devsuperior/dscatalog/resources/ProductResourceTest.java
package com.devsuperior.dscatalog.resources;

import com.devsuperior.dscatalog.dto.ProductDto;
import com.devsuperior.dscatalog.services.ProductService;
import com.devsuperior.dscatalog.services.exceptions.ResourceNotFoundException;
import com.fasterxml.jackson.databind.ObjectMapper;
import org.junit.jupiter.api.BeforeEach;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
import org.springframework.boot.test.mock.mockito.MockBean;
import org.springframework.data.domain.PageImpl;
import org.springframework.http.MediaType;
import org.springframework.test.web.servlet.MockMvc;
import org.springframework.test.web.servlet.ResultActions;

import java.util.List;

import static org.mockito.ArgumentMatchers.any;
import static org.mockito.ArgumentMatchers.eq;
import static org.mockito.Mockito.*;
import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;

@WebMvcTest(ProductResource.class)
public class ProductResourceTest {

    @Autowired
    private MockMvc mockMvc;

    @Autowired
    private ObjectMapper objectMapper;

    @MockBean
    private ProductService service;

    private PageImpl<ProductDto> page;

    private Long existingId;
    private Long nonExistingId;

    @BeforeEach
    void setUp() throws Exception {
        existingId = 1L;
        nonExistingId = -1000L;

        ProductDto productDto = new ProductDto();
        productDto.setId(1L);
        productDto.setName("Test");
        productDto.setDescription("Test");
        productDto.setPrice(10D);

        page = new PageImpl<>(List.of(productDto));
        when(service.findAllPaged(any())).thenReturn(page);

        when(service.findById(existingId)).thenReturn(productDto);
        when(service.findById(nonExistingId)).thenThrow(ResourceNotFoundException.class);

        doNothing().when(service).update(eq(existingId), any());
        doNothing().when(service).delete(eq(existingId));

        doThrow(ResourceNotFoundException.class).when(service).update(eq(nonExistingId), any());
        doThrow(ResourceNotFoundException.class).when(service).delete(eq(nonExistingId));

    }

    @Test
    public void deleteShouldReturnNoContentIdExists() throws Exception {
        var result = mockMvc.perform(delete("/products/{id}", existingId)
                .accept(MediaType.APPLICATION_JSON));
        result.andExpect(status().isNoContent());
    }

    @Test
    public void deleteShouldResourceNotFoundExceptionWhenNotIdExists() throws Exception {
        var result = mockMvc.perform(delete("/products/{id}", nonExistingId)
                .accept(MediaType.APPLICATION_JSON));

        result.andExpect(status().isNotFound());
    }

    @Test
    public void updateShouldReturnDTOWhenIdExists() throws Exception {
        ProductDto productDto = new ProductDto();
        productDto.setName("Test2");
        productDto.setDescription("TestDesc2");

        String jsonBody = objectMapper.writeValueAsString(productDto);
        var result = mockMvc.perform(put("/products/{id}", existingId)
                .content(jsonBody)
                .accept(MediaType.APPLICATION_JSON)
                .contentType(MediaType.APPLICATION_JSON));


        result.andExpect(status().isNoContent());
    }

    @Test
    public void updateShouldResourceNotFoundExceptionWhenNotIdExists() throws Exception {
        ProductDto productDto = new ProductDto();
        productDto.setName("Test2");
        productDto.setDescription("TestDesc2");

        String jsonBody = objectMapper.writeValueAsString(productDto);

        var result = mockMvc.perform(put("/products/{id}", nonExistingId)
                .accept(MediaType.APPLICATION_JSON)
                .contentType(MediaType.APPLICATION_JSON)
                .content(jsonBody));

        result.andExpect(status().isNotFound());
    }

    @Test
    public void findAllShouldReturnPage() throws Exception {
        ResultActions result = mockMvc.perform(get("/products")
                .accept(MediaType.APPLICATION_JSON));

        result.andExpect(status().isOk());
    }

    @Test
    public void findByIdShouldReturnProductDto() throws Exception {
        ResultActions result = mockMvc.perform(get("/products/{id}", existingId)
                .accept(MediaType.APPLICATION_JSON));

        result.andExpect(status().isOk());
        result.andExpect(jsonPath("$.id").exists());
        result.andExpect(jsonPath("$.name").exists());
        result.andExpect(jsonPath("$.description").exists());
    }

    @Test
    public void findByIdShouldReturnResourceNotFoundException() throws Exception {
        mockMvc.perform(get("/products/{id}", nonExistingId)
                .accept(MediaType.APPLICATION_JSON))
                .andExpect(status().isNotFound());
    }
}
